requirements.txt
aiohttp==3.8.6
requests==2.31.0
python-dotenv==1.0.0